+++ src/Ext/Techno/Hooks.cpp
@@ -359,6 +359,43 @@ DEFINE_HOOK(0x71067B, TechnoClass_EnterTransport_LaserTrails, 0x7)
 			pLaserTrail->Visible = false;
 			pLaserTrail->LastLocation = { };
 		}
+
+		if (pTechno->WhatAmI() == AbstractType::Infantry)
+		{
+			auto const pInf = abstract_cast<InfantryClass*>(pTechno);
+			auto const pInfType = abstract_cast<InfantryTypeClass*>(pTechno->GetTechnoType());
+
+			if (pInfType->Cyborg && pInf->Crawling)
+				pTechnoExt->IsLeggedCyborg = true;
+		}
+
+		if (pTechno->Transporter)
+		{
+			if (auto const pTransportTypeExt = TechnoTypeExt::ExtMap.Find(pTechno->Transporter->GetTechnoType()))
+			{
+				if (pTransportTypeExt->CanRepairCyborgLegs)
+					pTechnoExt->IsLeggedCyborg = false;
+			}
+		}
+	}
+
+	return 0;
+}
+
+DEFINE_HOOK(0x518047, TechnoClass_Destroyed_IsCyborg, 0x5)
+{
+	GET(InfantryClass*, pInf, ESI);
+	GET(DamageState, eDamageState, EAX);
+
+	if (pInf && eDamageState != DamageState::PostMortem)
+	{
+		if (pInf->Type->Cyborg)
+		{
+			auto pTechnoExt = TechnoExt::ExtMap.Find(pInf);
+
+			if (pTechnoExt && pInf->Type->Cyborg && pInf->Crawling == true && !pTechnoExt->IsLeggedCyborg)
+				pTechnoExt->IsLeggedCyborg = true;
+		}
 	}
 
 	return 0;
@@ -374,6 +411,15 @@ DEFINE_HOOK(0x6F6CFE, TechnoClass_Unlimbo_LaserTrails, 0x6)
 		{
 			pLaserTrail->LastLocation = { };
 			pLaserTrail->Visible = true;
+		}
+
+		// Fix legless Cyborgs when leave transports
+		if (pTechnoExt->IsLeggedCyborg)
+		{
+			InfantryClass* soldier = abstract_cast<InfantryClass*>(pTechno);
+			
+			soldier->SequenceAnim = Sequence::Prone;
+			soldier->Crawling = true;
 		}
 	}